% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{formatPc6}
\alias{formatPc6}
\title{Format an string or vec of strings as PC6.}
\usage{
formatPc6(postcodes)
}
\arguments{
\item{postcodes}{string or vec containing strings To be formatted as postcodes}
}
\value{
string or vec of srings Containing postcode6 formatted postal addresses.
}
\description{
This function attempts to format a given string or vec of strings as Dutch postal
addresses i.e. postcode6. Which is 4 numbers followed by two uppercase letters:
"1234AB". The function removes extra spaces from the given postal codes and makes
the letters uppercase. If it finds extra numbers or letters it errors.
}
