% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{loadElapse}
\alias{loadElapse}
\title{Load ELAPSE form given file or extdata.}
\usage{
loadElapse(path)
}
\arguments{
\item{path}{\emph{optional} Path to .tif image containing ELAPSE}
}
\value{
terra spatraster
}
\description{
Expects ELAPSE as a single tiff grey-scale image containing multiple layers.
When path argument is not given loads ELAPSE from inst/ELAPSE.tif otherwise
uses the given path.
}
