% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{pollutionFromBag}
\alias{pollutionFromBag}
\title{Estimate air pollution concentrations in a postcode with BAG database.}
\usage{
pollutionFromBag(postcodes, bag, elapse, ...)
}
\arguments{
\item{postcodes}{string or vector of strings Containing Duch postcodes.}

\item{bag}{string Path to BAG database.}

\item{elapse}{\emph{optional} Path to ELAPSE. (When empty will load internal ELAPSE)}
}
\value{
data.frame containing air pollution estimates and geo-location for all
buildings in the given postcodes
}
\description{
Kadaster's \href{https://www.pdok.nl/introductie/-/article/basisregistratie-adressen-en-gebouwen-ba-1}{Basisregistratie Adressen en Gebouwen}
database contains for all buildings in the Netherlands their location and postcode.
\code{pollutionFromBag()} finds the geo-location of all buildings part of the given
postcodes. Then extracts the air pollutant concentrations at those geo-locations
from ELAPSE. Returned are all pollution estimates for every building in the
wanted postcodes, including the geo-location.
}
\examples{
\dontrun{
# Basic usage
pollutionFromBag("9726AC", "bag-light.gpkg")

# Plotting the location of all buildings,
# I recommend you use tidyverse & tidyterra
library(tidyverse)
library(tidyterra)
library(postcodElapse)

data <- pollutionFromBag("9726AC", "bag-light.gpkg")

elapse <- loadElapse()

ggplot() +
   geom_spatraster(data = elapse$NO2FULL) +
   geom_sf(data = data, aes(geometry = geom), colour = "red")
}

}
