% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{pollutionFromPc6}
\alias{pollutionFromPc6}
\title{Get air pollution data for a postcode with Postcode6 database.}
\usage{
pollutionFromPc6(postcodes, pc6_path, elapse_path, ...)
}
\arguments{
\item{postcodes}{string or vector of strings Containing PC6 postcodes.}

\item{pc6_path}{string Path to the Postcode6 database.}

\item{elapse_path}{\emph{optional} string Path to the ELAPSE database.}
}
\value{
data.frame containing avrage, minimum & maximum air pollutant concentrations and areas per postcode.
}
\description{
Get air pollution data for a postcode with Postcode6 database.
}
