% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{coord_zoomFeature}
\alias{coord_zoomFeature}
\title{Zoom into given feature}
\usage{
coord_zoomFeature(data, id, r = 2000, ...)
}
\arguments{
\item{data}{data to zoom on.}

\item{id}{id of the data to zoom on}

\item{r}{radius of the zoom.}
}
\description{
Zoom into given feature
}
