% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{coord_zoomFeature}
\alias{coord_zoomFeature}
\title{Zoom into given feature}
\usage{
coord_zoomFeature(data, r = 2000, ...)
}
\arguments{
\item{data}{data to zoom on.}

\item{r}{radius of the zoom.}
}
\description{
Takes an simple features collection, i.e. the output form the polltionFrom*()
functions. Computes the center of the given features and builds a coordinate
transform that zooms into the given features.
}
\section{IMPORTANT NOTE}{

When passing a specific feature ensure you index using \verb{[2, ]}. Add the comma
otherwise R does not pass the entire row. \code{coord_zoomFeature()} expects the
entire row!
}

\examples{
\dontrun{
# You need to use this with a plot so the basics
library(tidyverse)
library(tidyterra)
library(postcodElapse)

data <- pollutionFromBag("9726AC", "bag-light.gpkg")

elapse <- loadElapse()

# We are zooming in on the 3rd building.
ggplot() +
   geom_spatraster(data = elapse$NO2FULL) +
   geom_sf(data = data, aes(geometry = geom), colour = "red") +
   coord_zoomFeature(data[3, ]) # you must index with `[3, ]` the entire row is expected!

# On the 5th trough 10th buildings
ggplot() +
   geom_spatraster(data = elapse$NO2FULL) +
   geom_sf(data = data, aes(geometry = geom), colour = "red") +
   coord_zoomFeature(data[5:10, ])
}

}
