% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{formatPc6}
\alias{formatPc6}
\title{Format strings as Dutch postcodes.}
\usage{
formatPc6(postcodes)
}
\arguments{
\item{postcodes}{Strings to be formatted as postcode}
}
\value{
Strings formatted as postcodes
}
\description{
This attempts to format any given string as an Dutch postcodes. IE. four numbers
followed by two capital letters, no spaces between each character. \code{formatPc6()}
removes any spaces in a given string and capitalizes all letters. In addition
extra numbers and or letters are checked for, resulting in an if they are present.
}
\examples{
formatPc6("8 9 3 3 DV", "9 713 Gz", "1071 xx")

}
