% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{geom_rectFeature}
\alias{geom_rectFeature}
\title{ggplot annotation that draws an rect around a given spacial feature.
Combine with \code{coord_zoomFeature()} for pretty insetting.}
\usage{
geom_rectFeature(data, id, size = 2000, colour = "red", fill = NA, ...)
}
\arguments{
\item{data}{data to zoom on.}

\item{id}{id of the data to zoom on}

\item{size}{size of the rectangle.}
}
\description{
ggplot annotation that draws an rect around a given spacial feature.
Combine with \code{coord_zoomFeature()} for pretty insetting.
}
