% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{geom_rectFeature}
\alias{geom_rectFeature}
\title{ggplot annotation that draws an rect around a given spacial feature.
Combine with \code{coord_zoomFeature()} for pretty insetting.}
\usage{
geom_rectFeature(data, size = 2000, colour = "red", fill = NA, ...)
}
\arguments{
\item{data}{data to zoom on.}

\item{size}{size of the square.}
}
\description{
ggplot annotation that draws an rect around a given spacial feature.
Combine with \code{coord_zoomFeature()} for pretty insetting.
}
\examples{
\dontrun{
# Some plotting basics
library(tidyverse)
library(tidyterra)
library(postcodElapse)

data <- pollutionFromBag("9726AC", "bag-light.gpkg")

elapse <- loadElapse()

# We are drawing a box around 3rd building.
ggplot() +
   geom_spatraster(data = elapse$NO2FULL) +
   geom_sf(data = data, aes(geometry = geom), colour = "red") +
   geom_reactFeature(data[3])

# On the 5th trough 10th buildings
ggplot() +
   geom_spatraster(data = elapse$NO2FULL) +
   geom_sf(data = data, aes(geometry = geom), colour = "red") +
   geom_rectFeature(data[5:10])
}

}
