% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{geom_rectFeature}
\alias{geom_rectFeature}
\title{Draws squares encompassing spacial features.}
\usage{
geom_rectFeature(data, size = 4000, colour = "red", fill = NA, ...)
}
\arguments{
\item{data}{data to zoom on.}

\item{size}{size of the rectangle.}

\item{colour}{border colour of the rectangle.}
}
\description{
Takes a collection of spacial features calculates the position and size of an
square encompassing all given spacial features. Builds and outputs ggplot
geom_rect to draw that square on a plot.
}
\section{IMPORTANT NOTE}{

When passing a specific feature ensure you index using \verb{[2, ]}. Add the comma
otherwise R does not pass the entire row. \code{geom_rectFeature()} expects the
entire row!
}

\examples{
\dontrun{
# Some plotting basics
library(tidyverse)
library(tidyterra)
library(postcodElapse)

data <- pollutionFromBag("9726AC", "bag-light.gpkg")

elapse <- loadElapse()

# Draw a box around 3rd building.
ggplot() +
   geom_spatraster(data = elapse$NO2FULL) +
   geom_sf(data = data, aes(geometry = geom), colour = "red") +
   geom_reactFeature(data[3, ])

# On the 5th trough 10th buildings
ggplot() +
   geom_spatraster(data = elapse$NO2FULL) +
   geom_sf(data = data, aes(geometry = geom), colour = "red") +
   geom_rectFeature(data[5:10, ])
}

}
