% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{loadElapse}
\alias{loadElapse}
\title{Load ELAPSE from given file or internal.}
\usage{
loadElapse(path)
}
\arguments{
\item{path}{\emph{optional} Path to .tif image containing ELAPSE}
}
\value{
terra spatraster
}
\description{
ELAPSE is a model representing air pollutant concentrations across Europe,
published by Kees de Hoogh. This function expects it as a tiff grey-scale image
with multiple layers per modeled air pollutant. When no file is given the internal
verion is loaded.
}
