% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{loadElapse}
\alias{loadElapse}
\title{Loads ELAPSE from the given path or from extdata.}
\usage{
loadElapse(path)
}
\arguments{
\item{path}{\emph{optional} Path to folder containing ELAPSE dataset.}
}
\value{
terra spatraster
}
\description{
Load the ELAPSE dataset, expected as a single .tif image containing all layers off ELAPSE.
When given no path the func loads the ELAPSE model stored in the package.
}
