% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{pollutionFromBag}
\alias{pollutionFromBag}
\title{Get air pollution data for a postcode with BAG database.}
\usage{
pollutionFromBag(postcodes, bag_path, elapse_path, ...)
}
\arguments{
\item{postcodes}{string or vector of strings Containing PC6 postcodes.}

\item{bag_path}{string Path to the BAG database.}

\item{elapse_path}{\emph{optional} string Path to the ELAPSE database.}
}
\description{
\code{pollution_from_bag()} extracts the geolocation of all building assigned the
given postcodes form BAG. With the geolocation the air pollution data can be
extracted from ELAPSE.
}
\section{BAG}{

Basisregistratie Adressen en Gebouwen or BAG contains the geolocation of all
buildings in the Netherlands with an address. You can download it from
\href{https://www.pdok.nl/introductie/-/article/basisregistratie-adressen-en-gebouwen-ba-1}{pdok}.
At the time of writing the file is 8gig.
}

\section{ELAPSE}{

This function automaticly loads the ELAPSE dataset included in the addrElapsR
package. When given elapse_path this func loads that ELAPSE dataset.
}

