% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{postcodElapse}
\alias{postcodElapse}
\title{Estimate air pollutant concentraions for a given Dutch postcode.}
\usage{
postcodElapse(postcodes, database_path, database_type = "GUESS", elapse_path)
}
\arguments{
\item{postcodes}{String containing Dutch postcode (1234AB) can also be a vector.}

\item{elapse_path}{When empty ELAPSE stored in extdata is used. Otherwise use
the one provided.}

\item{database}{Path to the postcode database, see Postcode database below.}

\item{type}{Type of the postcode database, defaults to "GUESS" where postcodElapse
fills type in as "BAG" for Basisregistratie Adressen en Gebouwen or
"PC6" for Postcode6.}
}
\description{
Estimates air pollutant concentration for given Dutch postcodes. Using ELAPSE
(included) and a postcode database (not-included). Two postcode databases formatted
as Geopackage (.gpkg) are supported \href{https://service.pdok.nl/cbs/postcode6/atom/postcode6_volledige_postcode.xml}{Postcode6(PC6)}
and \href{https://service.pdok.nl/lv/bag/atom/bag.xml}{Basisregistratie Adressen en Gebouwen(BAG)}
download one of these and point the \code{database } argument to it's location.
PostcodElapse automatically guesses the \code{type} argument, which you can set to
"PC6" or "BAG".
}
\examples{
\dontrun{
# Estimate air pollution for postcode "8933DV" with Postcode6.
postcodElapse("8933DV", "cbs_pc6_2024.gpkg")

# Estimate air pollution for multiple postcodes with BAG.
postcodElapse(c("8933DV", "9713 GZ", "1071XX"), "bag-light.gpkg")

# Note postcodes are valid a long as they are four numbers followed by two letters.
# In addition any spaces will be removed and non-capital letters will be capitalized.
postcodElapse(c("8 9 3 3 DV", "9 713 Gz", "1071 xx"), "cbs_pc6_2024.gpkg")
}
}
